attendance_management/attendance_management/attendance_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
openpyxl